requirements.txt
faiss_cpu==1.12.0
Flask==3.1.2
flask_cors==6.0.1
numpy==2.3.2
protobuf==6.32.0
sentence_transformers==5.1.0
tqdm==4.67.1
